+++ tests/lib/forms-schema.test.js
@@ -2727,6 +2727,126 @@ describe('CP_PAY submission event', () => {
   })
 })
 
+describe('CP_HCMS submission event', () => {
+  test('should allow CP_HCMS submission event', () => {
+    const { error } = Joi.validate(
+      {
+        id: 1,
+        name: 'string',
+        description: 'string',
+        formsAppEnvironmentId: 1,
+        formsAppIds: [1],
+        organisationId: 'ORGANISATION_00000000001',
+        postSubmissionAction: 'FORMS_LIBRARY',
+        isMultiPage: false,
+        elements: [
+          {
+            id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+            name: 'Numbers',
+            label: 'Numbers',
+            type: 'number',
+            required: false,
+          },
+        ],
+        isAuthenticated: true,
+        isPublished: true,
+        submissionEvents: [
+          {
+            type: 'CP_HCMS',
+            configuration: {
+              contentTypeName: 'contenttypename-1',
+            },
+          },
+        ],
+      },
+      formSchema,
+      {
+        abortEarly: false,
+      },
+    )
+    expect(error).toBeNull()
+  })
+  test('should reject CP_HCMS submission event with contentType name > 40', () => {
+    const { error } = Joi.validate(
+      {
+        id: 1,
+        name: 'string',
+        description: 'string',
+        formsAppEnvironmentId: 1,
+        formsAppIds: [1],
+        organisationId: 'ORGANISATION_00000000001',
+        postSubmissionAction: 'FORMS_LIBRARY',
+        isMultiPage: false,
+        elements: [
+          {
+            id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+            name: 'Numbers',
+            label: 'Numbers',
+            type: 'number',
+            required: false,
+          },
+        ],
+        isAuthenticated: true,
+        isPublished: true,
+        submissionEvents: [
+          {
+            type: 'CP_HCMS',
+            configuration: {
+              contentTypeName:
+                'contenttypename-1asasasasasasasasasasasasaasasasasa',
+            },
+          },
+        ],
+      },
+      formSchema,
+      {
+        abortEarly: false,
+      },
+    )
+    expect(error.message).toContain(
+      '"contentTypeName" length must be less than or equal to 40 characters long',
+    )
+  })
+  test('should reject CP_HCMS submission event with contentType name uppercase', () => {
+    const { error } = Joi.validate(
+      {
+        id: 1,
+        name: 'string',
+        description: 'string',
+        formsAppEnvironmentId: 1,
+        formsAppIds: [1],
+        organisationId: 'ORGANISATION_00000000001',
+        postSubmissionAction: 'FORMS_LIBRARY',
+        isMultiPage: false,
+        elements: [
+          {
+            id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+            name: 'Numbers',
+            label: 'Numbers',
+            type: 'number',
+            required: false,
+          },
+        ],
+        isAuthenticated: true,
+        isPublished: true,
+        submissionEvents: [
+          {
+            type: 'CP_HCMS',
+            configuration: {
+              contentTypeName: 'YELLINGNAME',
+            },
+          },
+        ],
+      },
+      formSchema,
+      {
+        abortEarly: false,
+      },
+    )
+    expect(error.message).toContain('fails to match the required pattern')
+  })
+})
+
 describe('Conditional Predicates', () => {
   test('should allow both OPTIONS and NUMERIC conditional types', () => {
     const { error } = Joi.validate(